This commit will added a new page to display all activities as the home page only display the Higher Todo activity priority to accomplish.
