+++ src/components/AddTodoForm/AddTodoForm.jsx
@@ -1,6 +1,5 @@
 import { useState } from "react";
 import InputWithLabel from "../InputWithLabel/InputWithLabel";
-import style from "./AddTodoForm.module.css"; // Import the CSS module
 import PropTypes from "prop-types";
 import Button from "react-bootstrap/Button";
 import Form from "react-bootstrap/Form";
